% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mondrian_forest.R
\name{convert_to_matrices}
\alias{convert_to_matrices}
\title{Creates matrices of covariates.}
\usage{
convert_to_matrices(X)
}
\arguments{
\item{X}{Data (matrix or data frame).}
}
\value{
A list of matrices for numeric and factor variables (if any).
}
\description{
Internal function to separate a dataset into matrices of numeric and factor variables.
}
\examples{
set.seed(1)
convert_to_matrices(data.frame(x1=rnorm(100),
                               x2=rbeta(100, 1, 2),
                               f1=sample(letters, 100, replace = TRUE)))
}
