% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_mondrian_forest.R
\name{extend_mondrian_forest}
\alias{extend_mondrian_forest}
\title{Extends a MondrianForest.}
\usage{
extend_mondrian_forest(mforest, newdata)
}
\arguments{
\item{mforest}{A mondrianforest.}

\item{newdata}{Data (matrix or data frame) containing new data to extend the
forest.}
}
\value{
A mondrianforest.
}
\description{
\code{mondrian_forest} implements a fast (looks at multiple new observations
at once) version of Lakshminarayanan et al's Extend Mondrian Block algorithm
described in [Lakshminarayanan et al. 2014]
(\url{https://arxiv.org/abs/1406.2673}) with a modification allowing for more
space-efficient treatment of categorical variables as if they were
dummy-encoded.
}
\examples{
library(mondrianforest); library(dplyr); library(purrr); library(magrittr)
set.seed(1)
test <- data.frame(x1 = rnorm(1000),
                   x2 = runif(1000),
                   x3 = rbeta(n = 1000, shape1 = 3, shape2 = 8),
                    # x3 is noise
                   x4 = rbinom(n = 1000, size = 4, prob = 0.2)) \%>\%
  map_df(function(x) (x - min(x))/(max(x) - min(x))) \%>\%
  mutate(y = x1*x2^2 + exp(x1) - x4,
         x1 = cut_number(sin(x1), n = 10),
         label = as.factor(case_when(y < 1 ~ "A",
                                     y >=1 & y < 1.7 ~ "B",
                                     y >= 1.7 ~ "C",
         ))) \%>\%
  select(-y)
mf <- mondrian_forest(test[1:5, ], y_col_num = 5, lambda = 3)
table(test$label[751:1000], predict(mf, test[751:1000, ], type = "class"))
mf_ext <- extend_mondrian_forest(mf, test[6:750, ])
table(test$label[751:1000], predict(mf_ext, test[751:1000, ], type = "class"))
}
