% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mondrian_forest.R
\name{predict.mondrianforest}
\alias{predict.mondrianforest}
\title{Predicts with an existing mondrianforest}
\usage{
\method{predict}{mondrianforest}(mforest, newdata, gamma = 1e-04,
  type = "prob")
}
\arguments{
\item{mforest}{A mondrianforest.}

\item{newdata}{A new dataset. If missing, predicts on training set.}

\item{gamma}{Hierarchical smoothing parameter, defaults to 0.0001.}
}
\value{
A matrix with class probabilities if type = "prob". If type = "class", returns MAP class estimates.
}
\description{
Predict method for object of class 'mondrianforest'.
}
\examples{
predict(mf, type = "class")

}
