% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.mondrianforest}
\alias{summary.mondrianforest}
\title{Summary method for mondrianforest.}
\usage{
\method{summary}{mondrianforest}(mforest)
}
\arguments{
\item{mforest}{a mondrianforest class object.}
}
\value{
A list containing the number of trees, dimensions of the training
  data, summary of the data, and parameters.
}
\description{
Return a list with a brief summary of the forest and training data.
}
